+++ src/abstractClasses/ProjectileTower.cpp
@@ -5,14 +5,14 @@ ProjectileTower::ProjectileTower(EnemyManager* enemyManager, Coords coords): Tow
     ProjectileTower::enemyManager = enemyManager;
     ProjectileTower::towerCoords = coords;
 
-    float exp = 0;
+    double exp = 0;
     for (int i = 0; i <= MAX_TOWER_LEVEL; i++){
         listExpForLvls.push_back(exp);
         exp += 100;
     }
 }
 
-ProjectileTower::ProjectileTower(EnemyManager* enemyManager, float damage, float radius, float attackSpeed, int level, Coords coords):
+ProjectileTower::ProjectileTower(EnemyManager* enemyManager, double damage, double radius, double attackSpeed, int level, Coords coords):
 Tower(enemyManager, damage, radius, attackSpeed, level, coords){
 
     ProjectileTower::enemyManager = enemyManager;
@@ -22,7 +22,7 @@ Tower(enemyManager, damage, radius, attackSpeed, level, coords){
     ProjectileTower::level = level;
     ProjectileTower::towerCoords = coords;
 
-    float exp = 0;
+    double exp = 0;
     for (int i = 0; i <= MAX_TOWER_LEVEL; i++){
         listExpForLvls.push_back(exp);
         exp += 100;
