+++ src/abstractClasses/Tower.h
@@ -22,11 +22,11 @@ using std::string;
 class Tower{
 public:
     Tower(EnemyManager* enemyManager, Coords coords);
-    Tower(EnemyManager* enemyManager, float damage, float radius, float attackSpeed, int level, Coords coords);
+    Tower(EnemyManager* enemyManager, double damage, double radius, double attackSpeed, int level, Coords coords);
     ~Tower();
 
-    float getDamage();
-    float getRadius();
+    double getDamage();
+    double getRadius();
     Coords getCoords();
 
     virtual void attack() = 0;
@@ -36,9 +36,9 @@ public:
 
     virtual void render(Renderer* renderer) = 0;
 
-    float getExpForNextLvl(int currentLevel);
+    double getExpForNextLvl(int currentLevel);
 
-    void setDamage(float damage);
+    void setDamage(double damage);
     void setCoords(Coords coords);
 
     bool isPointInRect(Coords point);
@@ -47,18 +47,18 @@ public:
     void setModeHoveredOver();
 
 protected:
-    float damage = 0;
-    float radius = 0;
+    double damage = 0;
+    double radius = 0;
 
     int level = 0;
-    float experience = 0;
-    float expForNextLvl = 0;
+    double experience = 0;
+    double expForNextLvl = 0;
 
-    float expForDamage = 0;
-    float expForKill = 0;
+    double expForDamage = 0;
+    double expForKill = 0;
 
     //in milliseconds
-    float attackSpeed = 0;
+    double attackSpeed = 0;
     
     Coords towerCoords;
 
@@ -77,20 +77,20 @@ protected:
 
     TexturesEnumeration towerTextureType;
 
-    std::vector<float> listExpForLvls;
+    std::vector<double> listExpForLvls;
 
     virtual void setAttackTimer() = 0;
     
-    void setExpForDamage(float exp);
-    void setExpForKill(float exp);
+    void setExpForDamage(double exp);
+    void setExpForKill(double exp);
 
     void findFirstEnemyInRadius();
     void findNearestEnemyInRadius();
 
-    void addExperience(float exp);
+    void addExperience(double exp);
     void checkAndLevelUp();
 
-    void setRadius(float radius);
+    void setRadius(double radius);
     void setLevel(int level);
 
     void renderRadiusCircle(Renderer* renderer);
